+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # PytSite Blog Changelog
 
+## 0.4 (2016-01-05)
+Update to support latest PytSite-0.41.17 release.
+
+
 ## 0.3 (2015-12-12)
-Updated to support latest PytSite-0.41.13 release.
+Update to support latest PytSite-0.41.13 release.
 
 
 ## 0.2 (2015-12-12)
-Updated to support latest PytSite-0.39 release.
+Update to support latest PytSite-0.39 release.
 
 
 ## 0.1 (2015-11-05)